+++ src/main/extensions/installer/semverCompare.js
