+++ src/main/extensions/installer/semverCompare.test.js
