+++ src/main/util/styling.js
@@ -1,1 +1,4 @@
 export const dialogBackgroundColor = "#d7dbe3";
+
+// these paths are relative to .webpack/main and not the src/
+export const assetPaths = ["css", "font-awesome-4.7.0"];
